DjangoRandomQuiz/requirements.txt
django==3.1
django-picklefield==3.0.1
